\name{add_datetime_fields}
\alias{add_datetime_fields}
\title{Adds datetime fields based on some of the common fields in the resources}
\usage{
  add_datetime_fields(entities)
}
\arguments{
  \item{entities}{representing instances of the resource to
  which we'd like to add datetime fields}
}
\description{
  Adds datetime fields based on some of the common fields
  in the resources
}

