\name{entity_list_to_data_frame}
\alias{entity_list_to_data_frame}
\title{Loads petition data from a JSON file.}
\usage{
  entity_list_to_data_frame(entity_list)
}
\arguments{
  \item{file}{the path to the JSON file, assumed to be in
  the format of a We The People API result}
}
\value{
  petitions a data frame of the petitions
}
\description{
  Loads petition data from a JSON file.
}

