\docType{data}
\name{IMPORTANT_DATES}
\alias{IMPORTANT_DATES}
\title{Some important dates relating to the We the People application and API.
Source: http://en.wikipedia.org/wiki/We_the_People_(petitioning_system)}
\format{List of 3
 $ application_launch      : POSIXct[1:1], format: "2011-09-22"
 $ threshold_raised_to_25k : POSIXct[1:1], format: "2011-10-03"
 $ threshold_raised_to_100k: POSIXct[1:1], format: "2013-01-16"}
\usage{
  IMPORTANT_DATES
}
\description{
  Some important dates relating to the We the People
  application and API. Source:
  http://en.wikipedia.org/wiki/We_the_People_(petitioning_system)
}
\keyword{datasets}

