\name{load_petitions}
\alias{load_petitions}
\title{Loads petition data from a JSON file.}
\usage{
  load_petitions(file)
}
\arguments{
  \item{file}{the path to the JSON file, assumed to be in
  the format of a We The People API result}
}
\value{
  petitions a data frame of the petitions
}
\description{
  Loads petition data from a JSON file.
}

