\name{melt_issues}
\alias{melt_issues}
\title{Melts the petitions data.frame, creating multiple rows for each petition issue.}
\usage{
  melt_issues(petitions)
}
\arguments{
  \item{petitions}{data.frame loaded using load_petitions}
}
\value{
  melted data.frame with one row per issue
}
\description{
  Melts the petitions data.frame, creating multiple rows
  for each petition issue.
}

