\name{petition_list_to_data_frame}
\alias{petition_list_to_data_frame}
\title{Converts a nested petition list to a data frame.}
\usage{
  petition_list_to_data_frame(petition_list)
}
\arguments{
  \item{petition_list}{a nested petition list}
}
\value{
  data frame of the entities
}
\description{
  Converts a nested petition list to a data frame.
}

