\name{petitions_from_json}
\alias{petitions_from_json}
\title{Transforms petitions from JSON to a data.frame}
\usage{
  petitions_from_json(petitions)
}
\arguments{
  \item{petitions}{nested lists from the json
  representation}
}
\value{
  users data.frame
}
\description{
  Transforms petitions from JSON to a data.frame
}

