\name{petitions_raw_to_data_frame}
\alias{petitions_raw_to_data_frame}
\title{Loads petition data from a JSON file.}
\usage{
  petitions_raw_to_data_frame(petitions_raw)
}
\arguments{
  \item{file}{the path to the JSON file, assumed to be in
  the format of a We The People API result}
}
\value{
  petitions a data frame of the petitions
}
\description{
  Loads petition data from a JSON file.
}

