\name{plot_issue_sentiment_bar}
\alias{plot_issue_sentiment_bar}
\title{Plots issue level sentiment in a simplistic bar chart.}
\usage{
  plot_issue_sentiment_bar(petitions)
}
\arguments{
  \item{petitions}{data.frame of petitions}
}
\description{
  Plots issue level sentiment in a simplistic bar chart.
}
\examples{
data(petitions)
plot_issue_sentiment_bar(petitions)
}

