\name{plot_issue_sentiment_boxplot}
\alias{plot_issue_sentiment_boxplot}
\title{Plots a boxplot of issues' sentiments.}
\usage{
  plot_issue_sentiment_boxplot(petitions)
}
\arguments{
  \item{petitions}{data.frame of petitions}
}
\description{
  Plots a boxplot of issues' sentiments.
}
\examples{
data(petition_analyses)
plot_issue_sentiment_boxplot(petition_analyses)
}

