\name{plot_issues_over_time}
\alias{plot_issues_over_time}
\title{Generates a plot of issues over time.}
\usage{
  plot_issues_over_time(petitions)
}
\arguments{
  \item{petitions}{a data frame of petitions}
}
\description{
  Generates a plot of issues over time.
}
\examples{
data(petitions)
plot_issues_over_time(petitions)
}

