\name{plot_issues_pending_response}
\alias{plot_issues_pending_response}
\title{Generates a plot of issues pending response over time.}
\usage{
  plot_issues_pending_response(petitions)
}
\arguments{
  \item{petitions}{a data frame of petitions}
}
\description{
  Generates a plot of issues pending response over time.
}
\examples{
data(petitions)
plot_issues_pending_response(petitions)
}

