\name{plot_sentiment_analysis}
\alias{plot_sentiment_analysis}
\title{Plots the sentiment together with petitions.}
\usage{
  plot_sentiment_analysis(petitions)
}
\arguments{
  \item{petitions}{data.frame of the petitions}
}
\description{
  Plots the sentiment together with petitions.
}
\examples{
data(petitions)
plot_sentiment_analysis(petitions)
}

