\name{plot_sentiment_density}
\alias{plot_sentiment_density}
\title{Plots the sentiment density.}
\usage{
  plot_sentiment_density(petitions)
}
\arguments{
  \item{petitions}{data.frame of the petitions}
}
\description{
  Plots the sentiment density.
}
\examples{
data(petitions)
plot_sentiment_density(petitions)
}

