\name{plot_signature_count}
\alias{plot_signature_count}
\title{Plots the signature counts over time.}
\usage{
  plot_signature_count(signatures)
}
\arguments{
  \item{signatures}{data.frame of the signatures to plot}
}
\value{
  ggplot2 plot object
}
\description{
  Plots the signature counts over time.
}

