\name{plot_status_by_creation}
\alias{plot_status_by_creation}
\title{Plots the statuses according to the creatio time.}
\usage{
  plot_status_by_creation(petitions)
}
\arguments{
  \item{petitions}{data.frame of petitions}
}
\value{
  plot of the status by creation date
}
\description{
  Plots the statuses according to the creatio time.
}

