\name{sample_petitions}
\alias{sample_petitions}
\title{Loads a small sample of petitions for use in examples.}
\usage{
  sample_petitions()
}
\value{
  a sample of 100 petitions
}
\description{
  Loads a small sample of petitions for use in examples.
}
\examples{
petitions <- sample_petitions()
}

