\name{signatures_from_json}
\alias{signatures_from_json}
\title{Transforms Signatures from JSON to a data.frame}
\usage{
  signatures_from_json(signatures)
}
\arguments{
  \item{signatures}{nested lists from the json
  representation}
}
\value{
  users data.frame
}
\description{
  Transforms Signatures from JSON to a data.frame
}

