\name{WeThePeopleMySQLClient}
\alias{WeThePeopleMySQLClient}
\title{We The People MySQL Client}
\usage{
  WeThePeopleMySQLClient(user, password, host, dbname)
}
\arguments{
  \item{user}{the MySQL username}

  \item{password}{the MySQL password}

  \item{host}{the MySQL host}

  \item{dbname}{the MySQL database name}
}
\value{
  WeThePeopleMySQLClient Object with methods for
  interfacing with the MySQL database instance
}
\description{
  We The People MySQL Client
}

